working on speed commands

+++ example/srcpy/wrappers.cpp
@@ -5,7 +5,7 @@ PYBIND11_MODULE(o80_example, m)
 {
     o80::create_python_bindings<o80_example::Driver,
                                 o80_example::Standalone,
-                                int,
-                                int>  // int,int are arguments to drivers
+                                double,
+                                double>  // int,int are arguments to drivers
         (m);
 }
